Rename Search to SearchInput
