+++ client/components/clubs/ClubList.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react'
 import styled from 'styled-components'
 import { ClubCard } from './ClubCard'
 import { PageTitle } from '../../styles'
-import { SearchInput } from './Search'
+import { SearchInput } from './SearchInput'
 
 const mobile = `425px`
 const tablet = `768px`
